v2/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
google-generativeai
pyairtable
pypdf
python-multipart
httpx

